+++ packages/prop-house-webapp/src/components/ProposalModalFooter/index.tsx
@@ -15,8 +15,7 @@ import ProposalModalVotingModule from '../ProposalModalVotingModule';
 import ProposalModalNavButtons from '../ProposalModalNavButtons';
 import VotesDisplay from '../VotesDisplay';
 import { useTranslation } from 'react-i18next';
-// import EditModeButtons from '../EditModeButtons';
-
+import ProposalWindowButtons from '../ProposalWindowButtons';
 
 const ProposalModalFooter: React.FC<{
   setShowVotingModal: Dispatch<SetStateAction<boolean>>;
